made migrations related to price, changed models with additionalPrice field

+++ modules/back/back.graphql.js
@@ -26,7 +26,7 @@ const backInputs = `
     model: ID
     features: BackFeatureSetInput
     image: Upload
-    additionalPrice: Int
+    additionalPrice: Float
     available: Boolean
     customizable: Boolean
   }
